Return modified/newly created object as responses
For now we're still returning only the current balance for the "deposit"
and "withdraw" operation as they can only modify one field of the
account. Returning the whole account in those operations felt too leaky
as the client could be gaining access without enough privilege should
such mechanisms be implemented.

Signed-off-by: Lucas Caparelli <lucas.caparelli112@gmail.com>